The details modal is a carousel and a panel, in separate files
graph-item-details-modal.tsx was 1809 lines and two components wearing one
hat: DetailsFilmstripModal, which lays a row of clips out and slides it,
and DetailsPanel, which is the whole details view for ONE clip. Nothing
was shared but a handful of constants, so the file was long rather than
cohesive.

Split along that seam. The modal keeps the carousel and the entry point;
the panel, the scoped history pair, the typed in/out fields and the
view-count arithmetic each become the module they always were. The few
values both halves need (the view-transition name, the panel gap, the
monitor sizes, the card lookup) go to a shared module so neither half
imports the other to agree on a number.

`rememberedViewCount` becomes lastViewCount()/rememberViewCount(): an ESM
import is a read-only binding, so a mutable `let` cannot be assigned
across a module edge — and naming the two operations is clearer than the
bare variable was anyway.

Behaviour-preserving by construction, and the modal's own 22 story
interactions pass unchanged, which is the point of doing it before the
redesign rather than inside it.
